refactor(eval-service): factorise le contexte vm sécurisé (secureContext.js)
Les protections de sécurité de l'évaluation étaient dupliquées et incohérentes
entre evalWorker.js et whereWorker.js (whereWorker n'avait pas importModuleDynamically).

Nouveau module partagé secureContext.js centralise :
- contexte vm neuf par job (import() dynamique rejeté via importModuleDynamically)
- libs épurées et gelées (lodash sans template/templateSettings — RCE host realm)
- stripDangerousGlobals (require/process/global/console/fetch/WebSocket)
- decodeUnicode

evalWorker.js et whereWorker.js utilisent désormais createSecureContext() :
protections HARMONISÉES entre eval et where (le where bénéficie maintenant du
blocage import()). Tests : pool 11, intégration 17. Perf ~4.5ms/éval.
